+++ index.html
@@ -6,7 +6,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>testpage</title>
-    <link rel="stylesheet" href="./main.css">
+    <link rel="stylesheet" href="./src/styles/main.css">
     <link rel="shortcut icon" href="#">
 </head>
 
@@ -89,7 +89,7 @@
         </div>
     </footer>
 
-    <script type="module" src="./main.js""></script>
+    <script type="module" src="./src/scripts/main.js""></script>
 </body>
 
 </html>
